fix: Address review comments
